Handle multiple methods with the same name in the target class
It needs to make sure that it gets the method with the same parameters and not just gets the first method with the same name. Because if it does that it might get the method with a different signature.

Given that the return type is not taken into the signature, but still needs to be equal, this has to be checked after a suitable method has been found.